requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.2
python-dotenv==1.0.1
boto3==1.34.162
Pillow==10.4.0
# mysqlclient==2.2.4  # 배포시 RDS(MySQL)용(로컬 SQLite는 주석 유지)
